requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
stripe==12.3.0
httpx==0.25.2
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
cryptography>=42.0.0
pandas==2.1.4
openpyxl==3.1.2
tabula-py==2.9.0
